Billing/requirements.txt
flask==3.0.2
mysql-connector-python==8.0.33
pandas
openpyxl